requirements.txt
# Core YOLOv11 and computer vision dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing and augmentation
albumentations>=1.3.0
imgaug>=0.4.0
scikit-learn>=1.3.0

# GUI and visualization
tkinter-tooltip>=2.0.0
customtkinter>=5.2.0
Pillow>=9.5.0

# Utilities and monitoring
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
psutil>=5.9.0
GPUtil>=1.4.0

# Configuration and logging
pyyaml>=6.0
loguru>=0.7.0
configparser>=5.3.0

# Additional utilities
pathlib2>=2.3.7
argparse>=1.4.0
